requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cymem==2.0.3
Cython==0.29.20
dartsclone==0.9.0
DAWG-Python==0.7.2
docopt==0.6.2
idna==2.9
importlib-metadata==1.6.1
jieba==0.42.1
joblib==0.15.1
more-itertools==8.4.0
murmurhash==1.0.2
natto-py==0.9.0
numpy==1.19.0
packaging==20.4
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
psutil==5.7.0
py==1.8.2
pycparser==2.20
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
pymorphy2-dicts-uk==2.4.1.1.1460299261
pyparsing==2.4.7
pytest==5.4.3
pytest-benchmark
pythainlp==2.2.0
python-crfsuite==0.9.7
pyvi==0.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn-crfsuite==0.3.6
sortedcontainers==2.1.0
spacy-lookups-data==0.3.2
srsly==1.0.2
SudachiDict-full==20201223.post1
SudachiPy==0.5.1
tabulate==0.8.7
thinc==7.4.1
threadpoolctl==2.1.0
tinydb==4.1.1
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.7.0
wcwidth==0.2.5
zipp==3.1.0
git+https://github.com/mikeizbicki/pymorphy2@ca1c13f#egg=pymorphy2
git+https://github.com/mikeizbicki/spacy@93da368#egg=spacy
lark==0.11.3
gensim==4.0.1
